+++ lib/geometry/Polygon.d.ts
@@ -79,6 +79,13 @@ export default class Polygon {
      */
     get position(): Vector;
     /**
+     * **Note:** Not sure if this will be kept or not but for now it's disabled.
+     *
+     * Sets a new position for this polygon and recalculates the points.
+     *
+     * @param {Vector} position A Vector representing the new position of this polygon.
+     */
+    /**
      * Returns the points of this polygon.
      *
      * @returns {Array<Vector>}
@@ -122,7 +129,7 @@ export default class Polygon {
      * just because of the inversion of the Y-axis when being displayed.
      *
      * @param {Array<Vector>} points An array of vectors representing the points in the polygon, in counter-clockwise order.
-     *
+     *    *
      * @returns {Polygon} Returns this for chaining.
      */
     setPoints(points: Array<Vector>): Polygon;