+++ src/geometry/Polygon.ts
@@ -96,6 +96,35 @@ export default class Polygon {
   get position(): Vector { return this._position; }
 
   /**
+   * **Note:** Not sure if this will be kept or not but for now it's disabled.
+   * 
+   * Sets a new position for this polygon and recalculates the points.
+   * 
+   * @param {Vector} position A Vector representing the new position of this polygon.
+   */
+  // set position(position: Vector) {
+  //   const diffX: number = -(this._position.x - position.x);
+  //   const diffY: number = -(this._position.y - position.y);
+
+  //   const diffPoint: Vector = new Vector(diffX, diffY);
+
+  //   const points: Array<Vector> = [];
+
+  //   this._points.map((point: Vector) => {
+  //     const tempX: number = point.x;
+  //     const tempY: number = point.y;
+
+  //     const tempPoint: Vector = new Vector(tempX, tempY);
+
+  //     const calculatedPoint: Vector = tempPoint.add(diffPoint);
+
+  //     points.push(calculatedPoint);
+  //   });
+
+  //   this.setPoints(points, true);
+  // }
+
+  /**
    * Returns the points of this polygon.
    * 
    * @returns {Array<Vector>}
@@ -145,7 +174,7 @@ export default class Polygon {
    * just because of the inversion of the Y-axis when being displayed.
    * 
    * @param {Array<Vector>} points An array of vectors representing the points in the polygon, in counter-clockwise order.
-   * 
+   *    * 
    * @returns {Polygon} Returns this for chaining.
    */
   setPoints(points: Array<Vector>): Polygon {