+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0 / 2020-02-18
+- [FEATURE] Added `translate` to circle to easily modify the position like the polygon version.
+- [TEST] Added testing for circle and polygon translations.
+
 ## 0.1.1 / 2020-02-18
 - [MISC] Replaced rollup-plugin-commonjs and rollup-plugin-node-resolve with @rollup/plugin-commonjs and @rollup/plugin-node-resolve.
 - [MISC] Updated dependencies to their latest versions.